chore: replace shared behavior test data
