+++ source/IntegrationTests/Behaviours/IntegrationEvents/GivenCalculationCompletedV1ReceivedForWholesaleFixingTests.cs
@@ -19,7 +19,6 @@ using Energinet.DataHub.EDI.B2BApi.Functions.EnqueueMessages.BRS_023_027.Activit
 using Energinet.DataHub.EDI.B2BApi.Functions.EnqueueMessages.BRS_023_027.Model;
 using Energinet.DataHub.EDI.BuildingBlocks.Domain.DataHub;
 using Energinet.DataHub.EDI.BuildingBlocks.Domain.Models;
-using Energinet.DataHub.EDI.IntegrationTests.Behaviours.IntegrationEvents.TestData;
 using Energinet.DataHub.EDI.IntegrationTests.Fixtures;
 using Energinet.DataHub.EDI.OutgoingMessages.Infrastructure.Databricks.WholesaleResults.Queries;
 using Energinet.DataHub.EDI.OutgoingMessages.Infrastructure.Extensions.Options;
