+++ source/IntegrationTests/Behaviours/TestData/ExampleEnergyResultMessageForActor.cs
@@ -13,20 +13,16 @@
 // limitations under the License.
 
 using Energinet.DataHub.EDI.BuildingBlocks.Domain.Models;
-using Energinet.DataHub.Edi.Responses;
-using Resolution = Energinet.DataHub.EDI.BuildingBlocks.Domain.Models.Resolution;
+using Energinet.DataHub.EDI.Tests.Infrastructure.OutgoingMessages.Asserts;
 
-namespace Energinet.DataHub.EDI.IntegrationTests.Behaviours.IntegrationEvents.TestData;
+namespace Energinet.DataHub.EDI.IntegrationTests.Behaviours.TestData;
 
-public record ExampleWholesaleResultMessageForActor(
+public record ExampleEnergyResultMessageForActor(
     string GridArea,
-    Currency Currency,
-    ActorNumber EnergySupplier,
-    MeteringPointType? MeteringPointType,
+    MeteringPointType MeteringPointType,
     SettlementMethod? SettlementMethod,
     Resolution Resolution,
+    ActorNumber? EnergySupplier,
+    ActorNumber? BalanceResponsible,
     int Version,
-    IReadOnlyCollection<WholesaleServicesRequestSeries.Types.Point> Points,
-    string? ChargeCode = null,
-    ChargeType? ChargeType = null,
-    MeasurementUnit? MeasurementUnit = null);
+    IReadOnlyCollection<TimeSeriesPointAssertionInput> Points);
