+++ source/IntegrationTests/Behaviours/TestData/wholesale_fixing_01-02-2023_28-02-2023_ga_803_and_804_amount_per_charge_v1.csv
