+++ source/IntegrationTests/Behaviours/TestData/wholesale_fixing_01-02-2023_28-02-2023_ga_804_total_amount_v1_with_invalid_row.csv
