+++ source/IntegrationTests/IntegrationTests.csproj
@@ -113,13 +113,13 @@ limitations under the License.
         <EmbeddedResource Include="Infrastructure.CimMessageAdapter\Messages\xml\RequestAggregatedMeasureData.xml">
             <CopyToOutputDirectory>Always</CopyToOutputDirectory>
         </EmbeddedResource>
-        <None Update="Behaviours\IntegrationEvents\TestData\balance_fixing_01-11-2022_01-12-2022_ga_542_and_543_per_brp_ga_v1.csv">
+        <None Update="Behaviours\TestData\balance_fixing_01-11-2022_01-12-2022_ga_542_and_543_per_brp_ga_v1.csv">
             <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
         </None>
-        <None Update="Behaviours\IntegrationEvents\TestData\balance_fixing_01-11-2022_01-12-2022_ga_542_and_543_per_es_brp_ga_v1.csv">
+        <None Update="Behaviours\TestData\balance_fixing_01-11-2022_01-12-2022_ga_542_and_543_per_es_brp_ga_v1.csv">
             <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
         </None>
-        <None Update="Behaviours\IntegrationEvents\TestData\balance_fixing_01-11-2022_01-12-2022_ga_543_per_ga_v1.csv">
+        <None Update="Behaviours\TestData\balance_fixing_01-11-2022_01-12-2022_ga_543_per_ga_v1.csv">
             <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
         </None>
         <None Update="appsettings.local.json">
@@ -128,31 +128,31 @@ limitations under the License.
         <None Update="integrationtest.local.settings.json">
             <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
         </None>
-        <None Update="Behaviours\IntegrationEvents\TestData\wholesale_fixing_01-02-2023_28-02-2023_ga_804_amount_per_charge_v1.csv">
+        <None Update="Behaviours\TestData\wholesale_fixing_01-02-2023_28-02-2023_ga_804_amount_per_charge_v1.csv">
             <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
         </None>
-        <None Update="Behaviours\IntegrationEvents\TestData\wholesale_fixing_01-02-2023_28-02-2023_ga_804_monthly_amount_per_charge_v1.csv">
+        <None Update="Behaviours\TestData\wholesale_fixing_01-02-2023_28-02-2023_ga_804_monthly_amount_per_charge_v1.csv">
             <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
         </None>
-        <None Update="Behaviours\IntegrationEvents\TestData\wholesale_fixing_01-02-2023_28-02-2023_ga_804_total_amount_v1.csv">
+        <None Update="Behaviours\TestData\wholesale_fixing_01-02-2023_28-02-2023_ga_804_total_amount_v1.csv">
             <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
         </None>
-        <None Update="Behaviours\IntegrationEvents\TestData\wholesale_fixing_01-02-2023_28-02-2023_ga_804_amount_per_charge_v1_with_invalid_row.csv">
+        <None Update="Behaviours\TestData\wholesale_fixing_01-02-2023_28-02-2023_ga_804_amount_per_charge_v1_with_invalid_row.csv">
             <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
         </None>
-        <None Update="Behaviours\IntegrationEvents\TestData\wholesale_fixing_01-02-2023_28-02-2023_ga_804_monthly_amount_per_charge_v1_with_invalid_row.csv">
+        <None Update="Behaviours\TestData\wholesale_fixing_01-02-2023_28-02-2023_ga_804_monthly_amount_per_charge_v1_with_invalid_row.csv">
             <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
         </None>
-        <None Update="Behaviours\IntegrationEvents\TestData\wholesale_fixing_01-02-2023_28-02-2023_ga_804_total_amount_v1_with_invalid_row.csv">
+        <None Update="Behaviours\TestData\wholesale_fixing_01-02-2023_28-02-2023_ga_804_total_amount_v1_with_invalid_row.csv">
             <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
         </None>
-        <None Update="Behaviours\IntegrationEvents\TestData\balance_fixing_01-11-2022_01-12-2022_ga_543_per_ga_v1_with_invalid_row.csv">
+        <None Update="Behaviours\TestData\balance_fixing_01-11-2022_01-12-2022_ga_543_per_ga_v1_with_invalid_row.csv">
             <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
         </None>
-        <None Update="Behaviours\IntegrationEvents\TestData\balance_fixing_01-11-2022_01-12-2022_ga_543_per_es_brp_ga_v1_with_invalid_row.csv">
+        <None Update="Behaviours\TestData\balance_fixing_01-11-2022_01-12-2022_ga_543_per_es_brp_ga_v1_with_invalid_row.csv">
             <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
         </None>
-        <None Update="Behaviours\IntegrationEvents\TestData\balance_fixing_01-11-2022_01-12-2022_ga_543_per_brp_ga_v1_with_invalid_row.csv">
+        <None Update="Behaviours\TestData\balance_fixing_01-11-2022_01-12-2022_ga_543_per_brp_ga_v1_with_invalid_row.csv">
             <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
         </None>
         <None Update="CalculationResults\TestData\wholesale_calculation_results.amounts_per_charge_v1.csv">
@@ -173,10 +173,10 @@ limitations under the License.
         <None Update="CalculationResults\TestData\wholesale_calculation_results.total_monthly_amounts_v1.csv">
             <CopyToOutputDirectory>Always</CopyToOutputDirectory>
         </None>
-        <None Update="Behaviours\IntegrationEvents\TestData\wholesale_fixing_01-02-2023_28-02-2023_ga_803_and_804_amount_per_charge_v1.csv">
+        <None Update="Behaviours\TestData\wholesale_fixing_01-02-2023_28-02-2023_ga_803_and_804_amount_per_charge_v1.csv">
             <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
         </None>
-        <None Update="Behaviours\IntegrationEvents\TestData\balance_fixing_01-11-2022_01-12-2022_ga_542_and_543_per_brp_ga_v1.csv">
+        <None Update="Behaviours\TestData\balance_fixing_01-11-2022_01-12-2022_ga_542_and_543_per_brp_ga_v1.csv">
             <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
         </None>
     </ItemGroup>
